# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sherpa MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/whatsapp/webhook": {
      "get": {
        "tags": [
          "whatsapp"
        ],
        "summary": "Verify Whatsapp",
        "description": "WhatsApp Cloud API Webhook verification.",
        "operationId": "verify_whatsapp_api_v1_whatsapp_webhook_get",
        "parameters": [
          {
            "name": "hub.mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Hub.Mode"
            }
          },
          {
            "name": "hub.verify_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Hub.Verify Token"
            }
          },
          {
            "name": "hub.challenge",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Hub.Challenge"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "whatsapp"
        ],
        "summary": "Whatsapp Webhook",
        "description": "Receive messages from WhatsApp Cloud API.",
        "operationId": "whatsapp_webhook_api_v1_whatsapp_webhook_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/v1/whatsapp/setup": {
      "post": {
        "tags": [
          "whatsapp"
        ],
        "summary": "Setup Whatsapp",
        "description": "Simplified Setup (Option B): Users only provide their number.\nThe Platform's master keys are used automatically.",
        "operationId": "setup_whatsapp_api_v1_whatsapp_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}